implement panning using a PIXI container's position
the same can be done with zoom (using container's scale)
but with 2x scale a 1px line will have 2px thickness, which is not
always what I want.

I like the way things look with current zoom, and I think I will have to
go through several cases to see if I want to redraw when zooming or not.
so I am leaving it the way it is. maybe at a later time I will decide to
implement it using container scale.