add gradient formation experiment
to make sure it works as described in the paper, I had to randomize the
timing of message sending, because in my simulation two neighbors might
happen to send their messages to a robot on the same tick, with no
loop() called in between. In the real world robots this might happen but
it is less likely. So, I randomized when a robot's message is sent by 1
tick to allow it to sometimes come before or after the one that would have
otherwise come exactlly on the same time.